[Feat] #65 Search Page added
